Motor Update 5
Added status frames to Spark Max, fully commented, Final Version?

+++ src/main/java/frc/team3128/common/hardware/motorcontroller/MotorControllerConstants.java
@@ -13,8 +13,8 @@ public class MotorControllerConstants {
     public static final double SPARKMAX_ENCODER_RESOLUTION = 42; // CPR
     public static final double SPARKMAX_RPM_TO_NUpS = SPARKMAX_ENCODER_RESOLUTION / 60; // rotations/min -> counts/sec
 
-    public static final double LOW_PRIORITY = 255;
-    public static final double MEDIUM_PRIORITY = 45;
-    public static final double HIGH_PRIORITY = 20;
-    public static final double ULTRA_PRIORITY = 10;
+    public static final int LOW_PRIORITY = 255;
+    public static final int MEDIUM_PRIORITY = 45;
+    public static final int HIGH_PRIORITY = 20;
+    public static final int ULTRA_PRIORITY = 10;
 }
